+++ tests/client.rs
@@ -1,5 +1,6 @@
 #![deny(warnings)]
 #![warn(rust_2018_idioms)]
+#![cfg_attr(feature = "deprecated", allow(deprecated))]
 
 #[macro_use]
 extern crate matches;
@@ -1121,10 +1122,11 @@ mod dispatch_impl {
     use http::Uri;
     use tokio::io::{AsyncRead, AsyncWrite, ReadBuf};
     use tokio::net::TcpStream;
+    use tokio_test::block_on;
 
     use super::support;
     use hyper::body::HttpBody;
-    use hyper::client::connect::{Connected, Connection, HttpConnector};
+    use hyper::client::connect::{capture_connection, Connected, Connection, HttpConnector};
     use hyper::Client;
 
     #[test]
@@ -1531,6 +1533,37 @@ mod dispatch_impl {
         // internal Connect::connect should have been lazy, and not
         // triggered an actual connect yet.
         assert_eq!(connects.load(Ordering::Relaxed), 0);
+    }
+
+    #[test]
+    fn capture_connection_on_client() {
+        let _ = pretty_env_logger::try_init();
+
+        let _rt = support::runtime();
+        let connector = DebugConnector::new();
+
+        let client = Client::builder().build(connector);
+
+        let server = TcpListener::bind("127.0.0.1:0").unwrap();
+        let addr = server.local_addr().unwrap();
+        thread::spawn(move || {
+            let mut sock = server.accept().unwrap().0;
+            //drop(server);
+            sock.set_read_timeout(Some(Duration::from_secs(5))).unwrap();
+            sock.set_write_timeout(Some(Duration::from_secs(5)))
+                .unwrap();
+            let mut buf = [0; 4096];
+            sock.read(&mut buf).expect("read 1");
+            sock.write_all(b"HTTP/1.1 200 OK\r\nContent-Length: 0\r\n\r\n")
+                .expect("write 1");
+        });
+        let mut req = Request::builder()
+            .uri(&*format!("http://{}/a", addr))
+            .body(Body::empty())
+            .unwrap();
+        let captured_conn = capture_connection(&mut req);
+        block_on(client.request(req)).expect("200 OK");
+        assert!(captured_conn.connection_metadata().is_some());
     }
 
     #[test]
@@ -2149,6 +2182,7 @@ mod dispatch_impl {
     }
 }
 
+#[allow(deprecated)]
 mod conn {
     use std::io::{self, Read, Write};
     use std::net::{SocketAddr, TcpListener};
@@ -2212,6 +2246,131 @@ mod conn {
         };
 
         future::join(server, client).await;
+    }
+
+    #[deny(deprecated)]
+    #[cfg(feature = "backports")]
+    mod backports {
+        use super::*;
+        #[tokio::test]
+        async fn get() {
+            let _ = ::pretty_env_logger::try_init();
+            let listener = TkTcpListener::bind(SocketAddr::from(([127, 0, 0, 1], 0)))
+                .await
+                .unwrap();
+            let addr = listener.local_addr().unwrap();
+
+            let server = async move {
+                let mut sock = listener.accept().await.unwrap().0;
+                let mut buf = [0; 4096];
+                let n = sock.read(&mut buf).await.expect("read 1");
+
+                // Notably:
+                // - Just a path, since just a path was set
+                // - No host, since no host was set
+                let expected = "GET /a HTTP/1.1\r\n\r\n";
+                assert_eq!(s(&buf[..n]), expected);
+
+                sock.write_all(b"HTTP/1.1 200 OK\r\nContent-Length: 0\r\n\r\n")
+                    .await
+                    .unwrap();
+            };
+
+            let client = async move {
+                let tcp = tcp_connect(&addr).await.expect("connect");
+                let (mut client, conn) = conn::http1::handshake(tcp).await.expect("handshake");
+
+                tokio::task::spawn(async move {
+                    conn.await.expect("http conn");
+                });
+
+                let req: Request<Body> = Request::builder()
+                    .uri("/a")
+                    .body(Default::default())
+                    .unwrap();
+                let mut res = client.send_request(req).await.expect("send_request");
+                assert_eq!(res.status(), hyper::StatusCode::OK);
+                assert!(res.body_mut().next().await.is_none());
+            };
+
+            future::join(server, client).await;
+        }
+
+        #[tokio::test]
+        async fn http2_detect_conn_eof() {
+            use futures_util::future;
+            use hyper::service::{make_service_fn, service_fn};
+            use hyper::{Response, Server};
+
+            let _ = pretty_env_logger::try_init();
+
+            let server = Server::bind(&([127, 0, 0, 1], 0).into())
+                .http2_only(true)
+                .serve(make_service_fn(|_| async move {
+                    Ok::<_, hyper::Error>(service_fn(|_req| {
+                        future::ok::<_, hyper::Error>(Response::new(Body::empty()))
+                    }))
+                }));
+            let addr = server.local_addr();
+            let (shdn_tx, shdn_rx) = oneshot::channel();
+            tokio::task::spawn(async move {
+                server
+                    .with_graceful_shutdown(async move {
+                        let _ = shdn_rx.await;
+                    })
+                    .await
+                    .expect("server")
+            });
+
+            struct TokioExec;
+            impl<F> hyper::rt::Executor<F> for TokioExec
+            where
+                F: std::future::Future + Send + 'static,
+                F::Output: Send + 'static,
+            {
+                fn execute(&self, fut: F) {
+                    tokio::spawn(fut);
+                }
+            }
+
+            let io = tcp_connect(&addr).await.expect("tcp connect");
+            let (mut client, conn) = conn::http2::Builder::new(TokioExec)
+                .handshake::<_, Body>(io)
+                .await
+                .expect("http handshake");
+
+            tokio::task::spawn(async move {
+                conn.await.expect("client conn");
+            });
+
+            // Sanity check that client is ready
+            future::poll_fn(|ctx| client.poll_ready(ctx))
+                .await
+                .expect("client poll ready sanity");
+
+            let req = Request::builder()
+                .uri(format!("http://{}/", addr))
+                .body(Body::empty())
+                .expect("request builder");
+
+            client.send_request(req).await.expect("req1 send");
+
+            // Sanity check that client is STILL ready
+            future::poll_fn(|ctx| client.poll_ready(ctx))
+                .await
+                .expect("client poll ready after");
+
+            // Trigger the server shutdown...
+            let _ = shdn_tx.send(());
+
+            // Allow time for graceful shutdown roundtrips...
+            tokio::time::sleep(Duration::from_millis(100)).await;
+
+            // After graceful shutdown roundtrips, the client should be closed...
+            future::poll_fn(|ctx| client.poll_ready(ctx))
+                .await
+                .expect_err("client should be closed");
+        }
     }
 
     #[tokio::test]
@@ -2993,6 +3152,61 @@ mod conn {
         future::poll_fn(|ctx| client.poll_ready(ctx))
             .await
             .expect("client should be open");
+    }
+
+    #[tokio::test]
+    async fn http2_responds_before_consuming_request_body() {
+        // Test that a early-response from server works correctly (request body wasn't fully consumed).
+        // https://github.com/hyperium/hyper/issues/2872
+        use hyper::service::service_fn;
+
+        let _ = pretty_env_logger::try_init();
+
+        let listener = TkTcpListener::bind(SocketAddr::from(([127, 0, 0, 1], 0)))
+            .await
+            .unwrap();
+        let addr = listener.local_addr().unwrap();
+
+        // Spawn an HTTP2 server that responds before reading the whole request body.
+        // It's normal case to decline the request due to headers or size of the body.
+        tokio::spawn(async move {
+            let sock = listener.accept().await.unwrap().0;
+            hyper::server::conn::Http::new()
+                .http2_only(true)
+                .serve_connection(
+                    sock,
+                    service_fn(|_req| async move {
+                        Ok::<_, hyper::Error>(http::Response::new(hyper::Body::from(
+                            "No bread for you!",
+                        )))
+                    }),
+                )
+                .await
+                .expect("serve_connection");
+        });
+
+        let io = tcp_connect(&addr).await.expect("tcp connect");
+        let (mut client, conn) = conn::Builder::new()
+            .http2_only(true)
+            .handshake::<_, Body>(io)
+            .await
+            .expect("http handshake");
+
+        tokio::spawn(async move {
+            conn.await.expect("client conn shouldn't error");
+        });
+
+        // Use a channel to keep request stream open
+        let (_tx, body) = hyper::Body::channel();
+        let req = Request::post("/a").body(body).unwrap();
+        let resp = client.send_request(req).await.expect("send_request");
+        assert!(resp.status().is_success());
+
+        let body = hyper::body::to_bytes(resp.into_body())
+            .await
+            .expect("get response body with no error");
+
+        assert_eq!(body.as_ref(), b"No bread for you!");
     }
 
     #[tokio::test]